Further Instrument class refactoring
