+++ WebApp/src/uk/ac/exeter/QuinCe/data/Instrument/Standards/StandardConcentration.java
@@ -1,9 +1,8 @@
-package uk.ac.exeter.QuinCe.data;
+package uk.ac.exeter.QuinCe.data.Instrument.Standards;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import uk.ac.exeter.QuinCe.data.Instrument.GasStandardDB;
 import uk.ac.exeter.QuinCe.data.Instrument.Instrument;
 import uk.ac.exeter.QuinCe.database.DatabaseException;
 import uk.ac.exeter.QuinCe.database.RecordNotFoundException;
