+++ WebApp/src/uk/ac/exeter/QuinCe/data/Instrument/Standards/StandardStub.java
@@ -1,4 +1,4 @@
-package uk.ac.exeter.QuinCe.data;
+package uk.ac.exeter.QuinCe.data.Instrument.Standards;
 
 import java.util.Date;
