+++ WebApp/src/uk/ac/exeter/QuinCe/database/Calculation/RawDataDB.java
@@ -21,13 +21,13 @@ import uk.ac.exeter.QCRoutines.config.RoutinesConfig;
 import uk.ac.exeter.QCRoutines.messages.Flag;
 import uk.ac.exeter.QuinCe.data.DateTimeParseException;
 import uk.ac.exeter.QuinCe.data.RawDataValues;
-import uk.ac.exeter.QuinCe.data.StandardConcentration;
-import uk.ac.exeter.QuinCe.data.StandardStub;
 import uk.ac.exeter.QuinCe.data.Calculation.GasStandardMean;
 import uk.ac.exeter.QuinCe.data.Calculation.GasStandardRuns;
-import uk.ac.exeter.QuinCe.data.Instrument.GasStandardDB;
 import uk.ac.exeter.QuinCe.data.Instrument.Instrument;
 import uk.ac.exeter.QuinCe.data.Instrument.InstrumentException;
+import uk.ac.exeter.QuinCe.data.Instrument.Standards.GasStandardDB;
+import uk.ac.exeter.QuinCe.data.Instrument.Standards.StandardConcentration;
+import uk.ac.exeter.QuinCe.data.Instrument.Standards.StandardStub;
 import uk.ac.exeter.QuinCe.database.DatabaseException;
 import uk.ac.exeter.QuinCe.database.DatabaseUtils;
 import uk.ac.exeter.QuinCe.database.RecordNotFoundException;
