+++ WebApp/src/uk/ac/exeter/QuinCe/jobs/files/DataReductionJob.java
@@ -20,10 +20,10 @@ import uk.ac.exeter.QCRoutines.messages.MissingValueMessage;
 import uk.ac.exeter.QuinCe.User.User;
 import uk.ac.exeter.QuinCe.data.FileInfo;
 import uk.ac.exeter.QuinCe.data.RawDataValues;
-import uk.ac.exeter.QuinCe.data.RunType;
 import uk.ac.exeter.QuinCe.data.Calculation.GasStandardRuns;
 import uk.ac.exeter.QuinCe.data.Instrument.Instrument;
 import uk.ac.exeter.QuinCe.data.Instrument.InstrumentDB;
+import uk.ac.exeter.QuinCe.data.Instrument.RunType;
 import uk.ac.exeter.QuinCe.data.QC.NoDataQCRecord;
 import uk.ac.exeter.QuinCe.data.QC.QCDB;
 import uk.ac.exeter.QuinCe.database.DatabaseException;
