+++ WebApp/src/uk/ac/exeter/QuinCe/jobs/files/ExtractRawDataJob.java
@@ -9,14 +9,14 @@ import java.util.Map;
 import java.util.Properties;
 
 import uk.ac.exeter.QuinCe.User.User;
-import uk.ac.exeter.QuinCe.data.CalibrationCoefficients;
-import uk.ac.exeter.QuinCe.data.CalibrationStub;
 import uk.ac.exeter.QuinCe.data.FileInfo;
 import uk.ac.exeter.QuinCe.data.RawDataFile;
-import uk.ac.exeter.QuinCe.data.SensorCode;
-import uk.ac.exeter.QuinCe.data.Instrument.CalibrationDB;
 import uk.ac.exeter.QuinCe.data.Instrument.Instrument;
 import uk.ac.exeter.QuinCe.data.Instrument.InstrumentDB;
+import uk.ac.exeter.QuinCe.data.Instrument.SensorCode;
+import uk.ac.exeter.QuinCe.data.Instrument.Calibration.CalibrationCoefficients;
+import uk.ac.exeter.QuinCe.data.Instrument.Calibration.CalibrationDB;
+import uk.ac.exeter.QuinCe.data.Instrument.Calibration.CalibrationStub;
 import uk.ac.exeter.QuinCe.data.QC.QCDB;
 import uk.ac.exeter.QuinCe.database.DatabaseException;
 import uk.ac.exeter.QuinCe.database.DatabaseUtils;
