+++ WebApp/src/uk/ac/exeter/QuinCe/web/Instrument/NewInstrumentBean.java
@@ -13,9 +13,9 @@ import java.util.TreeSet;
 
 import javax.annotation.PostConstruct;
 
-import uk.ac.exeter.QuinCe.data.RunType;
 import uk.ac.exeter.QuinCe.data.Instrument.Instrument;
 import uk.ac.exeter.QuinCe.data.Instrument.InstrumentDB;
+import uk.ac.exeter.QuinCe.data.Instrument.RunType;
 import uk.ac.exeter.QuinCe.database.DatabaseException;
 import uk.ac.exeter.QuinCe.utils.MissingParamException;
 import uk.ac.exeter.QuinCe.utils.StringUtils;
